pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "engram"
version = "0.1.0"
description = "Engram: Biologically-inspired memory layer for AI agents with forgetting, echo encoding, and dynamic categories"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Engram Team"}
]
keywords = ["memory", "ai", "agents", "forgetting", "llm", "echo", "categories"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0",
]

[project.optional-dependencies]
gemini = ["google-generativeai>=0.3.0"]
openai = ["openai>=1.0.0"]
qdrant = ["qdrant-client>=1.7.0"]
mcp = [
    "mcp>=1.0.0",
]
all = [
    "google-generativeai>=0.3.0",
    "openai>=1.0.0",
    "qdrant-client>=1.7.0",
    "requests>=2.28.0",
    "mcp>=1.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
engram-mcp = "engram.mcp_server:run"

[project.urls]
Homepage = "https://github.com/engram/engram"

[tool.setuptools.packages.find]
where = ["."]
include = ["engram*"]
